terminal_client/requirements.txt
websockets
sounddevice
numpy
pynput
loguru
pyperclip
python-dotenv
keyring
colorama
requests
